% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_health.R
\name{summary_health}
\alias{summary_health}
\title{Summary of Health Data}
\usage{
summary_health(df, x)
}
\arguments{
\item{df}{the data frame you want a summary of}

\item{x}{represents the categorical column in the data set the data will be grouped by.}
}
\value{
A data frame with a column with the name of the summary statistic, and a column with the resulting numbers for each level.
}
\description{
Prints the health summary statistics of the data set grouped by the given parameter.
}
\details{
For each categorical level the number of trees for each status is calculated. Additionally the mean and median percentage of crown intact and living is calculated.
}
\examples{
summary_health(data21, "crown_position")
summary_health(data22, "wounded_main_stem")

}
